Change post CRUD functions based on the user

+++ app/Http/Controllers/PostController.php
@@ -11,9 +11,14 @@ use App\Http\Resources\PostCollection;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Repositories\PostRepository;
+use App\Traits\HttpResponses;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
+
+    use HttpResponses;
+
     /**
      * Display a listing of the resource.
      * @param Request $request
@@ -28,21 +33,30 @@ class PostController extends Controller
     }
 
     /**
+     * Fetch posts by user
+     * @return PostCollection
+     */
+    public function myPost()
+    {
+        $myPosts = Post::where('user_id', Auth::user()->id)->get();
+        return new PostCollection($myPosts);
+    }
+
+    /**
      * Store a newly created resource in storage.
      *
      * @param  \App\Http\Requests\StorePostRequest  $request
      * @return PostResource
      */
-    public function store(StorePostRequest $request,PostRepository $repository)
+    public function store(StorePostRequest $request, PostRepository $repository)
     {
         $created = $repository->create($request->only([
             'title',
             'body',
-            'user_ids'
+            'category_id'
         ]));
 
         return new PostResource($created);
-
     }
 
     /**
@@ -63,12 +77,16 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return PostResource|JsonResponse
      */
-    public function update(UpdatePostRequest $request, Post $post,PostRepository $repository)
+    public function update(UpdatePostRequest $request, Post $post, PostRepository $repository)
     {
+        if ($this->isNotAuthorized($post)) {
+            return $this->isNotAuthorized($post);
+        }
+
         $post = $repository->update($post, $request->only([
             'title',
             'body',
-            'user_ids',
+            'category_id'
         ]));
 
         return new PostResource($post);
@@ -80,12 +98,30 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\JsonResponse
      */
-    public function destroy(Post $post,PostRepository $repository)
+    public function destroy(Post $post, PostRepository $repository)
     {
+        if ($this->isNotAuthorized($post)) {
+            return $this->isNotAuthorized($post);
+        }
+
         $post = $repository->forceDelete($post);
 
         return new JsonResponse([
             "data" => "success"
         ]);
+    }
+
+    /**
+     * Authorize user to perform certain action.
+     *
+     * @param  \App\Models\Post  $post
+     * @return JsonResponse
+     */
+    private function isNotAuthorized($post)
+    {
+
+        if (Auth::user()->id !== $post->user_id) {
+            return $this->error('', 'You are not authorized to make this request', 403);
+        }
     }
 }
